refactor(collector): split fetcher and feedWriter from collector index (#400)
- collectTypes.ts: CollectErrorKind / CollectResult / CollectAllResult 型と classifyError を集約
- fetcher.ts: fetchFeedOnce / fetchFeed / isRetryableError を移動
- feedWriter.ts: handleNotModified / handleSuccess / handleError と collectFeed を切り出し
- index.ts: 薄いオーケストレーター + re-export に整理 (627行 → 271行)
- feedWriter.test.ts: 3 ブランチの単体テストを追加 (14 ケース)

行数変化: index.ts 627 → 271行、新規 fetcher 136行、feedWriter 239行、collectTypes 59行
subrequest 数変更なし (fetch パターンは変更していない)
既存テスト (retry / conditionalGet / retention 等) 全通過
