+++ apps/web/worker/collector/index.ts
@@ -1,401 +1,27 @@
-import type { Env, FeedConfig } from "../types";
+import type { Env } from "../types";
 import { loadEnabledFeeds } from "../feed-config";
-import { parseFeed } from "./rssParser";
 import { parseXml, pickText, asArray } from "../utils/xml";
 import { checkUrlSafety } from "./url-safety";
-import { buildGuids } from "./deduplicator";
-import { D1CostAccumulator, writeCollectorEvent, writeD1CostEvent } from "./metrics";
+import { D1CostAccumulator, writeD1CostEvent } from "./metrics";
 import { maybeAlert, sendAlert } from "./alert";
 import {
-  buildInsertArticleStmts,
-  findExistingArticleGuids,
-  type InsertableArticle,
-} from "../db/articles";
-import {
-  buildRecordFetchErrorStmt,
-  buildRecordFetchSuccessStmt,
-  buildUpdateFeedHeadersStmt,
-  type FeedHeaders,
   getFeedStreaks,
   loadEnabledFeedIdsAndHeaders,
   syncFeeds,
+  type FeedHeaders,
 } from "../db/feeds";
-import { buildRecordRunFeedStmt, finishRun, startRun } from "../db/runs";
+import { finishRun, startRun } from "../db/runs";
+import {
+  classifyError,
+  type CollectAllResult,
+  type CollectErrorKind,
+  type CollectResult,
+} from "./collectTypes";
+import { fetchFeed, isRetryableError } from "./fetcher";
+import { collectFeed } from "./feedWriter";
 
-const MAX_FEED_BYTES = 4 * 1024 * 1024; // 4MB
-const MAX_ITEMS_PER_FEED = 50;
-const ALLOWED_URL_PREFIXES = ["http://", "https://"];
-// backoff: 500ms, 1000ms の 2 回まで。並列度 4 × max 1.5s = 6s < 30s cron 制限
-const BACKOFF_BASE_MS = 500;
-
-/**
- * エラーの種別。alert や監視ダッシュボードでエラーをフィルタリングできるよう型で区別する。
- * - timeout: AbortError (fetchFeed のタイムアウト)
- * - network: TypeError (DNS 解決失敗・接続拒否など)
- * - http_client: HTTP 4xx (404, 403 など恒久エラー)
- * - http_server: HTTP 5xx / 429 (一時障害。リトライ後も失敗した場合)
- * - parse: XML パース失敗または空フィード
- * - unknown: 上記に分類できないエラー
- */
-export type CollectErrorKind =
-  | "timeout"
-  | "network"
-  | "http_client"
-  | "http_server"
-  | "parse"
-  | "unknown";
-
-/** discriminated union: status に応じて error / errorKind フィールドの有無が変わる */
-export type CollectResult =
-  | { feedId: string; status: "ok"; inserted: number; parsed: number }
-  | { feedId: string; status: "not_modified"; inserted: number; parsed: number }
-  | {
-      feedId: string;
-      status: "error";
-      inserted: number;
-      parsed: number;
-      error: string;
-      errorKind: CollectErrorKind;
-    };
-
-/**
- * 例外からエラー種別を分類する。
- * isRetryableError と対になる分類で、同じ判定ロジックを使う。
- */
-export function classifyError(err: unknown): CollectErrorKind {
-  if (!(err instanceof Error)) return "unknown";
-  if (err.name === "AbortError") return "timeout";
-  if (err instanceof TypeError) return "network";
-  const m = /^HTTP (\d+)/.exec(err.message);
-  if (m) {
-    const code = Number(m[1]);
-    if (code === 429 || (code >= 500 && code < 600)) return "http_server";
-    if (code >= 400 && code < 500) return "http_client";
-  }
-  return "unknown";
-}
-
-export interface CollectAllResult {
-  total: number;
-  inserted: number;
-  pruned: number;
-  results: CollectResult[];
-  durationMs: number;
-}
-
-// 一部のブログ (mercari-engineering 等) は WAF が "bot" 名を含む UA を 403 で弾くため、
-// Mozilla 互換プレフィックスを付ける。Feedly / NewsBlur など主要 RSS リーダも同様の手法。
-const USER_AGENT =
-  "Mozilla/5.0 (compatible; tech-news-bot/0.1; +https://github.com/rikeda71/tech-news-bot)";
-
-/** 一時障害とみなしてリトライすべき HTTP ステータスかどうか */
-function isTransientStatus(status: number): boolean {
-  return status === 429 || (status >= 500 && status < 600);
-}
-
-/**
- * リトライ対象かどうかを判定する。
- * - AbortError: タイムアウト
- * - TypeError: ネットワーク接続失敗 ("Failed to fetch" 等)
- * - HTTP 5xx / 429: 一時的なサーバー障害
- * - HTTP 4xx (429 除く): 恒久エラーのためリトライしない
- * テスト用に export する
- */
-export function isRetryableError(err: unknown): boolean {
-  if (!(err instanceof Error)) return false;
-  if (err.name === "AbortError") return true;
-  if (err instanceof TypeError) return true;
-  const m = /^HTTP (\d+)/.exec(err.message);
-  if (m) return isTransientStatus(Number(m[1]));
-  // HTTP エラー以外の予期しないエラーはリトライしない
-  return false;
-}
-
-/** fetchFeedOnce の戻り値。304 の場合は xml が null になる。 */
-interface FetchFeedResult {
-  xml: string | null;
-  etag: string | null;
-  lastModified: string | null;
-  notModified: boolean;
-}
-
-/**
- * 1 回の HTTP fetch を試みる。失敗時は呼び出し元でリトライを判断する。
- * - 4xx (429 除く) は恒久エラーなので Error をそのまま throw
- * - 5xx / 429 / AbortError / ネットワークエラーは throw して呼び出し元がリトライ
- * - 304 Not Modified は notModified=true で返す (リトライしない)
- */
-async function fetchFeedOnce(
-  url: string,
-  timeoutMs: number,
-  conditionalHeaders: { etag: string | null; lastModified: string | null },
-): Promise<FetchFeedResult> {
-  const controller = new AbortController();
-  const timer = setTimeout(() => controller.abort(), timeoutMs);
-  try {
-    const reqHeaders: Record<string, string> = {
-      "User-Agent": USER_AGENT,
-      Accept: "application/atom+xml, application/rss+xml, application/xml;q=0.9, */*;q=0.5",
-    };
-    if (conditionalHeaders.etag) {
-      reqHeaders["if-none-match"] = conditionalHeaders.etag;
-    }
-    if (conditionalHeaders.lastModified) {
-      reqHeaders["if-modified-since"] = conditionalHeaders.lastModified;
-    }
-
-    const res = await fetch(url, {
-      headers: reqHeaders,
-      signal: controller.signal,
-      redirect: "follow",
-    });
-
-    // 304: サーバが変更なしと判断。parse/insert をスキップする。
-    if (res.status === 304) {
-      return { xml: null, etag: null, lastModified: null, notModified: true };
-    }
-
-    if (!res.ok) {
-      throw new Error(`HTTP ${res.status} ${res.statusText}`);
-    }
-
-    const contentLength = Number(res.headers.get("content-length") ?? 0);
-    if (contentLength > MAX_FEED_BYTES) {
-      throw new Error(`Feed too large: ${contentLength} bytes`);
-    }
-    const text = await res.text();
-    if (text.length > MAX_FEED_BYTES) {
-      throw new Error(`Feed body too large: ${text.length} bytes`);
-    }
-
-    return {
-      xml: text,
-      etag: res.headers.get("etag"),
-      lastModified: res.headers.get("last-modified"),
-      notModified: false,
-    };
-  } finally {
-    clearTimeout(timer);
-  }
-}
-
-/**
- * 一時障害 (5xx / 429 / AbortError / ネットワークエラー) に対して
- * 指数バックオフ + jitter でリトライする。4xx / 304 は即 return。
- * sleep は wallclock のみ消費するため Worker の CPU 制限に影響しない。
- * テスト用に export する
- */
-export async function fetchFeed(
-  url: string,
-  timeoutMs: number,
-  maxRetries: number,
-  conditionalHeaders: { etag: string | null; lastModified: string | null } = {
-    etag: null,
-    lastModified: null,
-  },
-): Promise<FetchFeedResult> {
-  let lastError: unknown;
-  for (let attempt = 0; attempt <= maxRetries; attempt++) {
-    try {
-      return await fetchFeedOnce(url, timeoutMs, conditionalHeaders);
-    } catch (err) {
-      lastError = err;
-
-      const isRetryable = isRetryableError(err);
-
-      if (!isRetryable || attempt >= maxRetries) break;
-
-      // base * 2^attempt + jitter(0..base)
-      const delayMs = BACKOFF_BASE_MS * 2 ** attempt + Math.random() * BACKOFF_BASE_MS;
-      console.warn(
-        `[collector] fetchFeed attempt ${attempt + 1} failed for ${url}: ${err instanceof Error ? err.message : String(err)}. Retrying in ${Math.round(delayMs)}ms`,
-      );
-      await new Promise<void>((resolve) => setTimeout(resolve, delayMs));
-    }
-  }
-  throw lastError;
-}
-
-function isSafeUrl(url: string): boolean {
-  return ALLOWED_URL_PREFIXES.some((p) => url.toLowerCase().startsWith(p));
-}
-
-/**
- * 1 feed 分の収集。per-feed の D1 書き込みを 1 batch にまとめて
- * Worker の subrequest 上限を圧迫しないようにする。
- *
- * Subrequest 内訳 (Cloudflare Workers の subrequest 上限対策):
- * - 304:    fetch (1) + batch[recordFetchSuccess + recordRunFeed?] (1) = 2
- * - 200:    fetch (1) + SELECT existing guids (1) + batch[updateHeaders + INSERTs + recordFetchSuccess + recordRunFeed?] (1) = 3
- * - error:  fetch (1) + batch[recordFetchError + recordRunFeed?] (1) = 2
- *
- * runId が null なら recordRunFeed は省略 (テスト等で run tracking 無効時)。
- */
-async function collectFeed(
-  env: Env,
-  feed: FeedConfig,
-  savedHeaders: FeedHeaders,
-  summaryMax: number,
-  timeoutMs: number,
-  maxRetries: number,
-  runId: number | null,
-  d1Acc: D1CostAccumulator,
-): Promise<CollectResult> {
-  const fetchedAt = new Date().toISOString();
-  const t0 = performance.now();
-  const feedStart = Date.now();
-  try {
-    // savedHeaders は collectAll 冒頭で全 enabled feed 分を 1 query でまとめて取得済み。
-    // feed あたり 1 read を消費しないので Worker の subrequest 上限対策になる。
-    const result = await fetchFeed(feed.url, timeoutMs, maxRetries, {
-      etag: savedHeaders.last_etag,
-      lastModified: savedHeaders.last_modified,
-    });
-
-    // 304 Not Modified: parse/insert をスキップし last_fetched_at だけ更新する
-    if (result.notModified) {
-      const stmts: D1PreparedStatement[] = [
-        buildRecordFetchSuccessStmt(env.DB, feed.id, fetchedAt, 0, "not_modified"),
-      ];
-      if (runId !== null) {
-        stmts.push(
-          buildRecordRunFeedStmt(env.DB, runId, feed.id, "skipped", 0, Date.now() - feedStart),
-        );
-      }
-      const batchResults = await env.DB.batch(stmts);
-      for (const r of batchResults) d1Acc.add(r.meta);
-      writeCollectorEvent(env, {
-        feedId: feed.id,
-        status: "not_modified",
-        ms: Math.round(performance.now() - t0),
-        statusCode: 304,
-      });
-      return { feedId: feed.id, status: "not_modified", inserted: 0, parsed: 0 };
-    }
-
-    // 200 OK path
-    const allItems = parseFeed(result.xml!, {
-      summaryMaxLength: summaryMax,
-      fallbackPublishedAt: fetchedAt,
-    });
-
-    const items = allItems.filter((i) => i.url && isSafeUrl(i.url)).slice(0, MAX_ITEMS_PER_FEED);
-
-    const guids = await buildGuids(
-      items.map((i) => ({
-        feedId: feed.id,
-        rawGuid: i.rawGuid,
-        url: i.url,
-        title: i.title,
-        publishedAt: i.publishedAt,
-      })),
-    );
-
-    const rows: InsertableArticle[] = items.map((item, idx) => ({
-      guid: guids[idx],
-      feed_id: feed.id,
-      title: item.title.slice(0, 500),
-      url: item.url!.slice(0, 1000),
-      summary: item.summary,
-      author: item.author ? item.author.slice(0, 200) : null,
-      published_at: item.publishedAt,
-      category: feed.category,
-      lang: feed.lang,
-    }));
-
-    // FTS トリガが INSERT OR IGNORE の changes() を変動させるため事前に既存 guid を引いて
-    // 新規行のみ INSERT する。SELECT は 1 subrequest だが per-feed batch とは独立して実行する。
-    const existingSet = await findExistingArticleGuids(
-      env.DB,
-      rows.map((r) => r.guid),
-    );
-    const newRows = rows.filter((r) => !existingSet.has(r.guid));
-    const inserted = newRows.length;
-
-    // header 更新 + INSERT 群 + recordFetchSuccess + recordRunFeed を 1 batch に集約。
-    // MAX_ITEMS_PER_FEED=50 + 3 = 53 stmts なので D1 batch 上限 100 を超えない。
-    const stmts: D1PreparedStatement[] = [
-      buildUpdateFeedHeadersStmt(env.DB, feed.id, result.etag, result.lastModified),
-      ...buildInsertArticleStmts(env.DB, newRows),
-      buildRecordFetchSuccessStmt(env.DB, feed.id, fetchedAt, inserted),
-    ];
-    if (runId !== null) {
-      stmts.push(
-        buildRecordRunFeedStmt(env.DB, runId, feed.id, "ok", inserted, Date.now() - feedStart),
-      );
-    }
-    const batchResults = await env.DB.batch(stmts);
-    for (const r of batchResults) d1Acc.add(r.meta);
-
-    writeCollectorEvent(env, {
-      feedId: feed.id,
-      status: "ok",
-      ms: Math.round(performance.now() - t0),
-      statusCode: 200,
-    });
-    return { feedId: feed.id, status: "ok", inserted, parsed: allItems.length };
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    const errorKind = classifyError(err);
-    try {
-      const stmts: D1PreparedStatement[] = [
-        buildRecordFetchErrorStmt(env.DB, feed.id, fetchedAt, message),
-      ];
-      if (runId !== null) {
-        stmts.push(
-          buildRecordRunFeedStmt(
-            env.DB,
-            runId,
-            feed.id,
-            "failed",
-            0,
-            Date.now() - feedStart,
-            message,
-          ),
-        );
-      }
-      const batchResults = await env.DB.batch(stmts);
-      for (const r of batchResults) d1Acc.add(r.meta);
-    } catch (logErr) {
-      console.error(`Failed to record error for ${feed.id}`, logErr);
-    }
-    // HTTP エラーコードをメッセージから抽出する (例: "HTTP 503 ...")
-    const statusMatch = /^HTTP (\d+)/.exec(message);
-    const statusCode = statusMatch ? Number(statusMatch[1]) : 0;
-    writeCollectorEvent(env, {
-      feedId: feed.id,
-      status: "error",
-      ms: Math.round(performance.now() - t0),
-      statusCode,
-    });
-    return {
-      feedId: feed.id,
-      status: "error",
-      inserted: 0,
-      parsed: 0,
-      error: message,
-      errorKind,
-    };
-  }
-}
-
-async function runWithConcurrency<T, R>(
-  items: T[],
-  concurrency: number,
-  worker: (item: T) => Promise<R>,
-): Promise<R[]> {
-  const results: R[] = Array.from({ length: items.length });
-  let cursor = 0;
-  const runners = Array.from({ length: Math.max(1, concurrency) }, async () => {
-    while (true) {
-      const idx = cursor++;
-      if (idx >= items.length) return;
-      results[idx] = await worker(items[idx]);
-    }
-  });
-  await Promise.all(runners);
-  return results;
-}
+export type { CollectErrorKind, CollectResult, CollectAllResult };
+export { classifyError, isRetryableError, fetchFeed };
 
 export type ValidateFeedResult =
   | { ok: true; title: string; lang: string | null; item_count: number }
@@ -474,6 +100,24 @@ export async function validateFeedUrl(url: string): Promise<ValidateFeedResult> 
   }
 
   return { ok: false, error: "not a valid RSS or Atom feed" };
+}
+
+async function runWithConcurrency<T, R>(
+  items: T[],
+  concurrency: number,
+  worker: (item: T) => Promise<R>,
+): Promise<R[]> {
+  const results: R[] = Array.from({ length: items.length });
+  let cursor = 0;
+  const runners = Array.from({ length: Math.max(1, concurrency) }, async () => {
+    while (true) {
+      const idx = cursor++;
+      if (idx >= items.length) return;
+      results[idx] = await worker(items[idx]);
+    }
+  });
+  await Promise.all(runners);
+  return results;
 }
 
 export async function collectAll(
